Int to Double changes

+++ fuzzy_octo_computing_machine/src/fuzzy_octo_computing_machine/Add_Sub.java
@@ -10,19 +10,19 @@ package fuzzy_octo_computing_machine;
  * @author matteo.lorenzini
  */
 public class Add_Sub {
-    int n1;
-    int n2;
+    double n1;
+    double n2;
     
-    public Add_Sub(int n1, int n2){
+    public Add_Sub(double n1, double n2){
         this.n1=n1;
         this.n2=n2;
     }
     
-    public int Add(int n1, int n2){
+    public double Add(double n1, double n2){
         return n1+n2;
     }
     
-    public int Sub(int n1, int n2){
+    public double Sub(double n1, double n2){
         return n1-n2;
     }
